finished

+++ event.h
@@ -19,7 +19,7 @@
             int getOccurChance();
             void setOccurChance(int);
             EVENT_TYPE getEventType();
-            int getLevel();
+            double getLevel();
     };
 
 #endif
